requirements.txt
azure-storage-blob==12.19.0
aspose-words==23.12.0
aspose-pdf==23.12.0
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0 